refactor(gc): split the sticky remembered-set buffer and the first-cycle wrapper out of copying.rs
