+++ crates/perry-runtime/src/gc/mod.rs
@@ -128,9 +128,13 @@ use pin::{note_preflight_skipped, note_preflight_walked, young_pin_latch_armed};
 mod prefetch;
 
 mod copying;
+mod copying_first_cycle;
 /// Per-scanner root attribution for the copied-minor root scan (#7915).
 mod scanner_profile;
+mod sticky_remembered;
 use copying::*;
+use copying_first_cycle::*;
+use sticky_remembered::*;
 // The copied-minor pointer classifier is consumed by the weak-holder registry
 // pass in `crate::weakref` (#6182), which lives outside the gc module.
 pub(crate) use copying::CopyingPointerSet;
